% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{create_checkbox_variable}
\alias{create_checkbox_variable}
\title{Create a combined checkbox variable from individual variables}
\usage{
create_checkbox_variable(project, variable, sep = "|")
}
\arguments{
\item{project}{A tibble of redcap data}

\item{variable}{A variable to create based on individual choices}

\item{sep}{Separator for individual choices in the variable}
}
\value{
A tibble column representing \code{variable} which combines values from checkbox variables.
}
\description{
Create a combined checkbox variable from individual variables
}
