% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{csv_filename}
\alias{csv_filename}
\title{Return the CSV filename based on input filename}
\usage{
csv_filename(f)
}
\arguments{
\item{f}{A REDCap filename (either csv or R)}
}
\value{
A CSV filename derived from the input filename.
}
\description{
Given a filename, return the R filename regardless
of which input (csv, R) is provided.
}
\examples{
csv_filename("PartnershipTrackingP_R_2021-11-08_1628.r")
}
