% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{csv_filename_to_r}
\alias{csv_filename_to_r}
\title{Transform CSV filename to R filename.}
\usage{
csv_filename_to_r(f)
}
\arguments{
\item{f}{A filename to transform.}
}
\value{
A R filename, equivalent to the CSV filename
}
\description{
Transform CSV filename to R filename.
}
\examples{
csv_filename_to_r("PartnershipTrackingP_DATA_2021-11-08_1628.csv")
}
