% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{Embed columns into a longer, nested tibble}
\usage{
embed(x, vars, target, embedded_names = NULL)
}
\arguments{
\item{x}{A data frame to embed variables within}

\item{vars}{The columns (variables) to extract and embed.}

\item{target}{The column name for the embedded data.}

\item{embedded_names}{(Default NULL) The embedded column name for variable names (if they are kept).
Use NULL to keep only column values, not the column name.}
}
\value{
A data frame with \code{vars} embedded as a tibble, with each variable in a separate row.
}
\description{
Embed a selected list of columns (variables) into a \code{long} tibble that
is nested per row.
}
\details{
This function provides a two-step transformation. When provided with a list of \code{vars},
it will extract these variables and transpose them into an embedded data frame with each row of
the data frame representing one of the variables in the original table.
Behind the scenes, it will use \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} to transform these variables into repeated rows.
It will then nest these repeated rows into a tibble.
}
\examples{
\dontrun{
embed(iris, c("Sepal.Length","Sepal.Width"), "Sepal.Stuff", embedded_names="Sepal.Variable")
}
}
