% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{extract_checkbox_choice}
\alias{extract_checkbox_choice}
\title{Extract checkbox choice from string}
\usage{
extract_checkbox_choice(s)
}
\arguments{
\item{s}{A string to extract the checkbox choice from.}
}
\value{
The choice string
}
\description{
A checkbox choice is embedded in a string describing the question.
}
\details{
The checkbox is embedded in a string of the type\preformatted{This is the question? (choice=<choice>)
}

This function extracts the \verb{<choice>} from the text.
}
\examples{
extract_checkbox_choice( "Which one or more of the following are blue? (choice=sky)")
# [1] "sky"
}
