% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{find_factor_pairs}
\alias{find_factor_pairs}
\title{Find variable pairs with factor and non-factor entries}
\usage{
find_factor_pairs(x)
}
\arguments{
\item{x}{A redcap instrument}
}
\value{
A tibble of pairs (\code{unlabelled} and \code{labelled}).
}
\description{
Find variable pairs with factor and non-factor entries
}
