% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{format_redcap_filename}
\alias{format_redcap_filename}
\title{Format REDCap components to a filename.}
\usage{
format_redcap_filename(comp)
}
\arguments{
\item{comp}{A list of components. See \code{\link{split_redcap_filename}}.}
}
\value{
A character vector filename corresponding to the components.
}
\description{
The components of a REDCap file, as a list of components, can
be reassembled into a filename string. This function does that.
}
\examples{
format_redcap_filename(list(preamble="PartnershipTracking",
type="DATA",
date="2021-11-08",
time = "1628",
ext="csv"
))

}
