% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from}
\alias{from}
\alias{\%from\%}
\title{Extract the names that are in a list}
\usage{
x \%from\% y
}
\arguments{
\item{x}{A target list or regex to extract}

\item{y}{A list of values to search against}
}
\value{
A list of elements within \code{y} that match \code{x}.
}
\description{
Extract a list (or regex) \code{x} from a list \code{y}, returning the actual values.
}
\details{
This may be a re-implementation of an existing function, however often it would be
helpful to extract specific elements and return these from within a list (possible with a
regex). This function does that.
}
\examples{
"A" \%from\% c("A","B","C")
}
