% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_dictionary.R
\name{import_data_dictionary}
\alias{import_data_dictionary}
\title{Import REDCap-exported data dictionary}
\usage{
import_data_dictionary(f)
}
\arguments{
\item{f}{A filename representing the data dictionary export.}
}
\value{
A tibble (see details) of information.
}
\description{
REDCap provides a csv-formatted data dictionary, describing variables,
instruments and the variable type. This function loads the dictionary into R.
}
\examples{
\dontrun{import_data_dictionary("data_dictionary_export_from_redcap.csv")}
}
