% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{is_filename_csv}
\alias{is_filename_csv}
\title{Is the filename a csv file?}
\usage{
is_filename_csv(f)
}
\arguments{
\item{f}{A filename to test}
}
\value{
Logical. If the filename is a csv file.
}
\description{
Is the filename a csv file?
}
\examples{
is_filename_csv("test.csv")
# [1] TRUE
}
