% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{is_filename_r}
\alias{is_filename_r}
\title{Is the filename an R script?}
\usage{
is_filename_r(f)
}
\arguments{
\item{f}{A filename to test}
}
\value{
Logical. If the filename is an r script.
}
\description{
Is the filename an R script?
}
\examples{
is_filename_r("test.R")
# [1] TRUE

}
