% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_data.R
\name{label_factor_variables}
\alias{label_factor_variables}
\title{Apply Label Information to Factor Variables in REDCap Output}
\usage{
label_factor_variables(x)
}
\arguments{
\item{x}{A table of REDCap imported data.}
}
\value{
A tibble of REDCap imported data with factor variables having labels.
}
\description{
By default, REDCap imported data does not include the label information
for the variables that are transformed to factors. This function adds them in.
}
\details{
The REDCap default import is great. It knows what variables should be factors and turns them
into factors (while keeping the original numerical forms). This is really helpful when summarizing
REDCap data. However, a minor point is that, although the numerical forms have labels (e.g., the
question asked), the converted factor variables do not.

This function adds in labels to the factor variables. It does this by simply copying over the
label that was on the non-factor version of the variable. The resulting data is then returned.
}
\note{
This function is not exported and should generally not be called outside of the package.
It is part of a set of function calls to clean up redcap data. Check out \code{\link{read_redcap_data}}
for the best entrypoint to this code.
}
