% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{or}
\alias{or}
\alias{\%||\%}
\title{Return a default value if missing}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Value to return, unless null}

\item{b}{Default value to return}
}
\value{
The parameter \code{a}, unless it is null, in which case \code{b} is returned.
}
\description{
If a value is null, return a default instead.
}
\examples{
NA \%||\% 1
}
