% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{r_filename}
\alias{r_filename}
\title{Return the R filename based on input filename}
\usage{
r_filename(f)
}
\arguments{
\item{f}{A REDCap filename (either csv or R)}
}
\value{
A R filename derived from the input filename.
}
\description{
Given a filename, return the R filename regardless
of which input (csv, R) is provided.
}
\examples{
r_filename("PartnershipTrackingP_DATA_2021-11-08_1628.csv")
# [1] "PartnershipTrackingP_R_2021-11-08_1628.r"
}
