% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{r_filename_to_csv}
\alias{r_filename_to_csv}
\title{Transform R filename to CSV filename.}
\usage{
r_filename_to_csv(f)
}
\arguments{
\item{f}{A filename to transform.}
}
\value{
A CSV filename, equivalent to the R filename
}
\description{
Transform R filename to CSV filename.
}
\examples{
r_filename_to_csv("PartnershipTrackingP_R_2021-11-08_1628.r")
}
