% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{recode_checkbox_variable}
\alias{recode_checkbox_variable}
\title{Recode the checkbox variable based on the value embedded in variable label}
\usage{
recode_checkbox_variable(variable)
}
\arguments{
\item{variable}{A labelled vector representing a checkbox variable.}
}
\value{
A checkbox
}
\description{
A checkbox variable consists of two components: a value that the checkbox
represents (in the variable label) and the status of the checkbox for a series of
observations. This function recodes the variable to the checkbox value.
}
\details{
A checkbox variable consists of two components: a value that the checkbox
represents (in the variable label) and the status of the checkbox for a series of
observations. This function will recode a variable (checkbox) into value/NA with the
value being the checkbox's value.

For example, consider an input variable\preformatted{checkbox_variable___1
"Which one or more of the following are blue? (choice=sky)"
[1] 0 0 0 0 1 0
}

Note that the question is the \code{\link[labelled]{var_label}} for the variable. The values
of 0/1 indicate that the choice was selected or not. The result of this function should
return\preformatted{"Which one or more of the following are blue? (choice=sky)"
 [1] NA NA NA NA "sky" NA
}
}
\examples{
recode_checkbox_variable(
 labelled::set_variable_labels(data.frame(sky_choice=c(0,0,0,0,1,0)),
               sky_choice="Which one or more of the following are blue? (choice=sky)")
)
#     sky_choice
#[1,] NA
#[2,] NA
#[3,] NA
#[4,] NA
#[5,] "sky"
#[6,] NA
}
