% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{return_filename_pair}
\alias{return_filename_pair}
\title{Return REDCap filename pairs (R and DATA)}
\usage{
return_filename_pair(f)
}
\arguments{
\item{f}{A filename to derive pairs from}
}
\value{
A list of two filenames: the R and DATA elements.
}
\description{
Return REDCap filename pairs (R and DATA)
}
\examples{
return_filename_pair("PartnershipTrackingP_DATA_2021-11-08_1628.csv")

}
