% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{split_instrument_by_labels}
\alias{split_instrument_by_labels}
\title{Split instrument into labelled and unlabelled versions}
\usage{
split_instrument_by_labels(x)
}
\arguments{
\item{x}{An input REDCap instrument (or raw combination of instruments).}
}
\value{
A list of two tibbles, \code{labelled} and \code{unlabelled}.
}
\description{
Split instrument into labelled and unlabelled versions
}
