% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_dictionary.R
\name{split_levels}
\alias{split_levels}
\title{Split a REDCap-encoded level into Value and Label}
\usage{
split_levels(s)
}
\arguments{
\item{s}{A string representing a level (vectorized).}
}
\value{
A tibble (\code{Value}, \code{Label}) representing the encoded string.
}
\description{
Split a string representing a REDCap-encoded level into two
components: a value and a label.
}
\details{
REDCap provides Choices in the form of \verb{2, Good}. This function will
split this encoding into a tibble of two columns called \code{Value} and \code{Label}.
}
\examples{

split_levels(c("1, Poor", "2, Good"))
}
