% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_project_into_instruments.R
\name{split_project_into_instruments}
\alias{split_project_into_instruments}
\title{Split project data into separate instruments}
\usage{
split_project_into_instruments(project, dictionary)
}
\arguments{
\item{project}{The REDCap tibble}

\item{dictionary}{A REDCap data dictionary tibble}
}
\value{
A list of tibbles
}
\description{
REDCap data is organized into different instruments which correspond to
different surveys or data collection tools. This function will use the data dictionary
to split the combined data from a REDCap project into different instruments.
}
