% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{split_redcap_filename}
\alias{split_redcap_filename}
\title{Split Filename into REDCap Components}
\usage{
split_redcap_filename(f)
}
\arguments{
\item{f}{The filename to split apart.}
}
\value{
A list of filename components, or the filename (plus a warning) in
the case of an error.
}
\description{
Split a filename from REDCap export into the components, as
a named list.
}
\details{
A REDCap export file set consists of two filenames with specific
format names. This function splits the name into components which are
then returned in a named list. The items of the list are
\describe{
\item{preamble}{The name of the file, without the REDCap pieces}
\item{type}{The type, which is either R or DATA}
\item{date}{The YYYY-MM-DD of the export}
\item{time}{The HHMM of the export}
\item{ext}{The file extensions}
}
}
\examples{
split_redcap_filename("PartnershipTrackingP_DATA_2021-11-08_1628.csv")
split_redcap_filename("PartnershipTrackingP_R_2021-11-08_1628.r")
}
