% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{coalesce_checkbox_fields}
\alias{coalesce_checkbox_fields}
\title{Coalesce checkbox variables into a single list}
\usage{
coalesce_checkbox_fields(project)
}
\arguments{
\item{project}{A redcap tibble}
}
\value{
A modified tibble with checkbox fields replaced with single columns
}
\description{
A checkbox variable in redcap comes in as yes/no on multiple fields. These
are coalesced into a single variable which is a list of selections.
}
